rename search icon jsx file to match current folder structure format
