+++ src/components/atoms/icons/SvgSearch/SvgSearch.js
