+++ src/components/atoms/icons/SvgSearch/SvgSearch.stories.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 
-import SvgSearch from './index';
+import SvgSearch from './SvgSearch';
 
 storiesOf('Atoms/Icons', module)
   .add('SVG Search', () => <SvgSearch />);
